fix(mcts): widen action_idx u16→u32 to support infinite board coordinates
The u16 encoding stored only 8 bits per coordinate (q+128, r+128) in high/low
byte, silently overflowing when |q| or |r| > 127. On long games the decoded
(q,r) pointed to already-occupied cells, triggering the 'selected move should
always be legal' panic in selection.rs.

Fix: encode each coordinate as 16-bit with offset 32768 packed into a u32.
Supports q,r in [-32768, 32767] — sufficient for any practical game length.
All decode sites in mod.rs, selection.rs and hardcoded test values updated.

Also: add ownership_loss and threat_loss to train_step JSONL log event;
reduce train.smoke from 200 → 20 steps for faster verification.

+++ engine/src/mcts/node.rs
@@ -17,7 +17,7 @@ pub struct TTEntry {
 #[derive(Clone, Copy)]
 pub struct Node {
     pub parent: u32,
-    pub action_idx: u16,
+    pub action_idx: u32,
     pub n_visits: u32,
     pub w_value: f32,
     pub prior: f32,
@@ -33,7 +33,7 @@ impl Node {
     pub fn uninit() -> Self {
         Node {
             parent: u32::MAX,
-            action_idx: u16::MAX,
+            action_idx: u32::MAX,
             n_visits: 0,
             w_value: 0.0,
             prior: 0.0,
